Configure and Route Database
Add CreateAcronym to the list of migrations to run
Set the log level for the application to debug. This provides more information and enables you to see your migrations.
Automatically run migrations and wait for the result. Fluent allows you to choose when to run you migrations. This is helpful when you need to schedule them, for example. You can use wait() here since you're not running on an EventLoop.
Define a new type, Create Acronym that conforms to Migration.
Implement prepare(on:) as required by Migration. You call this method when you run your migrations.
Define the table name for this model. This must match schema from the model.
Define the ID column in the database.
Define columns for short and long. Set the column type to string and mark the columns as required. This matches the non-optional String properties in the model. The field names must match the key of the propery wrapper.
Implement revert(on:) as required by Migration. You call this function when you revert your migrations. This deletes the table referenced with schema(_:).
Define a Class that conforms to Model.
Specify the schema as required by Model. This is the name of the table in the database.
Define two String properties to hold the acronym and its definition. These use the @Field property wrapper to denote a generic database field. The key parameter is the name of the column in the database.
Define two String properties to hold the acronym and its definition. These use the @Field property wrapper to denote a generic database field. The key parameter is the name of the column in the database.
Provide an empty initializer as required by Model. Fluent uses this to initialize models returned from database queries.
Provide an initializer to create the model.
Register a new route at /api/acronyms that accepts a POST request and returns EventLoopFuture<Acronym>. It returns the acronym once it's saved.
Decode the requests's JSON into an Acronym model using Codable.
Save the model using Fluent and the databse from Requests
save(on:) returns EventLoopFuture<Void> so use map to return the acronym when the save completes.